requirement.txt
asgiref @ file:///C:/ci_311/asgiref_1676448296195/work
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django @ file:///C:/ci_311/django_1676469589007/work
django-allauth==0.54.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
idna==3.4
oauthlib==3.2.2
Pillow==10.0.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
pymongo==3.12.1
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.2.4
tzdata @ file:///tmp/build/80754af9/python-tzdata_1611693927741/work
urllib3==2.0.2
